+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/StructureIndexProvider.java
@@ -1,6 +1,6 @@
 package org.rcsb.strucmotif.io;
 
-import org.rcsb.strucmotif.domain.query.ContentType;
+import org.rcsb.strucmotif.domain.query.ResultsContentType;
 
 import java.util.Collection;
 import java.util.Set;
@@ -53,8 +53,8 @@ public interface StructureIndexProvider {
 
     /**
      * Get all registered indices of a particular type.
-     * @param contentTypes search space identifier (PDB entries, computed structure models, everything)
+     * @param resultsContentType search space identifier (PDB entries, computed structure models, everything)
      * @return a collection of all relevant structure indices
      */
-    Set<Integer> selectByContentTypes(Collection<ContentType> contentTypes);
+    Set<Integer> selectByResultsContentType(Collection<ResultsContentType> resultsContentType);
 }