+++ strucmotif-search-core/src/test/java/org/rcsb/strucmotif/domain/query/ResultsContentTypeTest.java
@@ -6,27 +6,27 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class SearchSpaceTest {
+class ResultsContentTypeTest {
     private static final Set<String> TEST_CASES = Set.of("1acj", "1MUW", "1eXr", "AF-Q8W3K0-F1", "ma-bak-cepc-0001");
 
     @Test
     public void whenHandlingPdbIds_thenArchiveEntriesRetained() {
         assertEquals(3, TEST_CASES.stream()
-                .filter(ContentType.EXPERIMENTAL)
+                .filter(ResultsContentType.EXPERIMENTAL)
                 .count());
     }
 
     @Test
     public void whenHandlingModelIds_thenArchiveEntriesIgnored() {
         assertEquals(2, TEST_CASES.stream()
-                .filter(ContentType.COMPUTATIONAL)
+                .filter(ResultsContentType.COMPUTATIONAL)
                 .count());
     }
 
     @Test
     public void whenAllSearchSpace_thenNoOperation() {
         assertEquals(TEST_CASES.size(), TEST_CASES.stream()
-                .filter(ContentType.EXPERIMENTAL.or(ContentType.COMPUTATIONAL))
+                .filter(ResultsContentType.EXPERIMENTAL.or(ResultsContentType.COMPUTATIONAL))
                 .count());
     }
 }